# [cherry-pick] [branch-2.2] [Enhancement] Optimize the memory usage of rows between unbounded preceding and current row (#5883)